Fix Unix builds
On Unix, the zapsig.h is somehow included before sigparser.h and thus
the usage of SigParser type in the zapsig.h fails.

+++ src/vm/zapsig.cpp
@@ -1121,7 +1121,7 @@ void ZapSig::CopyTypeSignature(SigParser* pSigParser, SigBuilder* pSigBuilder, D
         CorElementType type = (CorElementType)byteType;
 
         // Some element types have sub-type that is what we are interested in
-        switch (type)
+        switch ((DWORD)type)
         {
             case ELEMENT_TYPE_BYREF:
             case ELEMENT_TYPE_PTR:
